+++ POSSystem/src/application/controller/SaleController.java
@@ -105,7 +105,7 @@ public class SaleController {
 	}
 	
 	public String getSaleNumber() {
-		return Integer.toString(newSale.getSaleNumber());
+		return Long.toString(newSale.getSaleNumber());
 	}
 	public String getSaleTotal() {
 		return newSale.getPrintSaleTotal();
